major updates to the widgets.  NetworkList Widget has new mouse handling to more efficiently trigger plot updates.  Invalid db values are no longer plotted, correct handling of a list of corrected ntwk values, refactored NetworkListWidget and NetworkPlotWidget into separate files, and other changes as well
