--- qtapps/skrf_qtwidgets/designer/measurement.ui
@@ -1,177 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MeasurementDialog</class>
- <widget class="QDialog" name="MeasurementDialog">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>547</width>
-    <height>359</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Dialog</string>
-  </property>
-  <layout class="QHBoxLayout" name="horizontalLayout_5">
-   <item>
-    <layout class="QVBoxLayout" name="verticalLayout_left">
-     <item>
-      <widget class="QGroupBox" name="groupBox_options">
-       <property name="title">
-        <string>Options</string>
-       </property>
-       <layout class="QVBoxLayout" name="verticalLayout_2">
-        <item>
-         <widget class="QCheckBox" name="checkBox_sweepNew">
-          <property name="text">
-           <string>Sweep New</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <layout class="QHBoxLayout" name="horizontalLayout_channel">
-          <item>
-           <widget class="QLabel" name="label_channel">
-            <property name="text">
-             <string>Channel</string>
-            </property>
-           </widget>
-          </item>
-          <item>
-           <widget class="QSpinBox" name="spinBox_channel"/>
-          </item>
-         </layout>
-        </item>
-       </layout>
-      </widget>
-     </item>
-     <item>
-      <widget class="QGroupBox" name="groupBox_snp">
-       <property name="title">
-        <string>Get N-Port Network</string>
-       </property>
-       <layout class="QVBoxLayout" name="verticalLayout">
-        <item>
-         <layout class="QHBoxLayout" name="horizontalLayout_nports">
-          <item>
-           <widget class="QLabel" name="label_nports">
-            <property name="text">
-             <string>NPorts:</string>
-            </property>
-           </widget>
-          </item>
-          <item>
-           <widget class="QComboBox" name="comboBox_nports">
-            <item>
-             <property name="text">
-              <string>1</string>
-             </property>
-            </item>
-            <item>
-             <property name="text">
-              <string>2</string>
-             </property>
-            </item>
-           </widget>
-          </item>
-         </layout>
-        </item>
-        <item>
-         <layout class="QHBoxLayout" name="horizontalLayout_port1">
-          <item>
-           <widget class="QLabel" name="label_port1">
-            <property name="text">
-             <string>Port1</string>
-            </property>
-           </widget>
-          </item>
-          <item>
-           <widget class="QSpinBox" name="spinBox_port1"/>
-          </item>
-         </layout>
-        </item>
-        <item>
-         <layout class="QHBoxLayout" name="horizontalLayout_port2">
-          <item>
-           <widget class="QLabel" name="label_port2">
-            <property name="text">
-             <string>Port2</string>
-            </property>
-           </widget>
-          </item>
-          <item>
-           <widget class="QSpinBox" name="spinBox_port2"/>
-          </item>
-         </layout>
-        </item>
-        <item>
-         <widget class="QPushButton" name="pushButton_measureSnp">
-          <property name="text">
-           <string>Measure Network</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <spacer name="verticalSpacer">
-          <property name="orientation">
-           <enum>Qt::Vertical</enum>
-          </property>
-          <property name="sizeHint" stdset="0">
-           <size>
-            <width>20</width>
-            <height>40</height>
-           </size>
-          </property>
-         </spacer>
-        </item>
-       </layout>
-       <zorder>label_nports</zorder>
-       <zorder>comboBox_nports</zorder>
-       <zorder></zorder>
-       <zorder></zorder>
-       <zorder>label_nports</zorder>
-       <zorder></zorder>
-       <zorder>comboBox_nports</zorder>
-       <zorder></zorder>
-       <zorder>verticalSpacer</zorder>
-       <zorder>pushButton_measureSnp</zorder>
-      </widget>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <widget class="QGroupBox" name="groupBox_traces">
-     <property name="title">
-      <string>Available Traces</string>
-     </property>
-     <layout class="QVBoxLayout" name="verticalLayout_3">
-      <item>
-       <widget class="QListWidget" name="listWidget_traces"/>
-      </item>
-      <item>
-       <layout class="QHBoxLayout" name="horizontalLayout_tracesButtons">
-        <item>
-         <widget class="QPushButton" name="btn_updateTraces">
-          <property name="text">
-           <string>Update</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QPushButton" name="btn_measureTraces">
-          <property name="text">
-           <string>Measure Traces</string>
-          </property>
-         </widget>
-        </item>
-       </layout>
-      </item>
-     </layout>
-    </widget>
-   </item>
-  </layout>
- </widget>
- <resources/>
- <connections/>
-</ui>
