--- qtapps/skrf_qtwidgets/designer/sparameters_plot.ui
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Form</class>
- <widget class="QWidget" name="Form">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>520</width>
-    <height>478</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <layout class="QVBoxLayout" name="verticalLayout">
-   <item>
-    <layout class="QHBoxLayout" name="horizontalLayout">
-     <item>
-      <widget class="QCheckBox" name="checkBox_useCorrected">
-       <property name="text">
-        <string>Plot Corrected</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QComboBox" name="comboBox_unitsSelector"/>
-     </item>
-     <item>
-      <widget class="QComboBox" name="comboBox_traceSelector"/>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <widget class="NetworkPlot" name="ntwkPlot">
-     <property name="sizePolicy">
-      <sizepolicy hsizetype="Expanding" vsizetype="Expanding">
-       <horstretch>10</horstretch>
-       <verstretch>0</verstretch>
-      </sizepolicy>
-     </property>
-    </widget>
-   </item>
-  </layout>
- </widget>
- <customwidgets>
-  <customwidget>
-   <class>NetworkPlot</class>
-   <extends>QGraphicsView</extends>
-   <header>pyMultiCal.widgets</header>
-  </customwidget>
- </customwidgets>
- <resources/>
- <connections/>
-</ui>
