+++ qtapps/skrf_qtwidgets/numeric_inputs.py
@@ -9,8 +9,8 @@ available_units = {
     },
     "length": {
         "base": "m",
-        "m": 1.0, "dm": 10, "cm": 100, "mm": 1e3, "km": 1e-3,
-        "in": 1000 / 25.4, "mil": 1e6 / 25.4
+        "m": 1.0, "dm": 10, "cm": 100, "mm": 1e3, "um": 1e6, "nm": 1e9, "pm": 1e12, "km": 1e-3,
+        "yd": 1000 / 25.4 / 36,  "ft": 1000 / 25.4 / 12, "in": 1000 / 25.4, "mil": 1e6 / 25.4
     }
 }
