+++ qtapps/skrf_qtwidgets/widgets.py
@@ -1,18 +1,12 @@
 import os
-import re
 import sys
 import traceback
-from collections import OrderedDict
-from math import sqrt
 
-import numpy as np
-import pyqtgraph as pg
 from qtpy import QtWidgets, QtCore, QtGui
 
 import skrf
 from . import numeric_inputs
 from . import qt
-from . import smith_chart
 from .analyzers import analyzers
 
 
@@ -51,29 +45,6 @@ def load_network_files(caption="load network file", filter="touchstone file (*.s
     return ntwks
 
 
-def trace_color_cycle(n=1000):
-    """
-    :type n: int
-    :return:
-    """
-
-    lime_green = "#00FF00"
-    cyan = "#00FFFF"
-    magenta = "#FF00FF"
-    yellow = "#FFFF00"
-    pink = "#C04040"
-    blue = "#0000FF"
-    lavendar = "#FF40FF"
-    turquoise = "#00FFFF"
-
-    count = 0
-    colors = [yellow, cyan, magenta, lime_green, pink, blue, lavendar, turquoise]
-    num = len(colors)
-    while count < n:
-        yield colors[count % num]
-        count += 1
-
-
 def save_multiple_networks(ntwk_list):
     dirname = qt.getDirName_Global("select directory to save network files")
     if not dirname:
@@ -106,65 +77,6 @@ def save_multiple_networks(ntwk_list):
                     fname = qt.getSaveFileName_Global("save network file", filter)
 
             ntwk.write_touchstone(fname)
-
-
-@QtCore.Slot(object, str)
-def save_NetworkListItem(ntwk_list_item, save_which):
-    """
-    :type ntwk_list_item: NetworkListItem
-    :type save_which: str
-    :return:
-    """
-
-    if save_which.lower() not in ("raw", "cal", "both"):
-        raise ValueError("Must set save option to 'raw', 'cal', or 'both'")
-
-    ntwk = ntwk_list_item.ntwk
-    ntwk_c = ntwk_list_item.ntwk_corrected
-
-    if type(ntwk) in (list, tuple):
-        save_multiple_networks(ntwk)
-        return
-    elif not isinstance(ntwk, skrf.Network):
-        raise TypeError("ntwk must be a skrf.Network object to save")
-
-    extension = ".s{:d}p".format(ntwk.s.shape[1])
-    file_filter = "touchstone format (*{:s})".format(extension)
-    filename = os.path.join(qt.cfg.last_path, ntwk.name + extension)
-
-    if save_which.lower() == "both":
-        if isinstance(ntwk, skrf.Network) and isinstance(ntwk_c, skrf.Network):
-            filename = qt.getSaveFileName_Global("Save Raw skrf.Network File", filter=file_filter, start_path=filename)
-            if not filename:
-                return
-            base, ext = os.path.splitext(filename)
-            filename_cal = base + "-cal" + ext
-            filename_cal = qt.getSaveFileName_Global("Save Calibrated skrf.Network File",
-                                                 filter=file_filter,
-                                                 start_path=filename_cal)
-            if not filename_cal:
-                return
-            ntwk.write_touchstone(filename)
-            ntwk_c.write_touchstone(filename_cal)
-            return
-        else:
-            save_which = "raw"
-
-    if save_which.lower() == "cal":
-        if ntwk_list_item.ntwk_corrected is None:
-            Warning("ntwk_corrected is None, saving raw instead")
-            save_which = "raw"
-        else:
-            ntwk = ntwk_list_item.ntwk_corrected
-            filename = os.path.join(qt.cfg.last_path, ntwk.name + extension)
-            if not filename:
-                return
-
-    caption = "Save skrf.Network File" if save_which.lower() == "raw" else "Save Calibrated skrf.Network File"
-    filename = qt.getSaveFileName_Global(caption, filter=file_filter, start_path=filename)
-    if not filename:
-        return
-    ntwk.write_touchstone(filename)
 
 
 class OverwriteFilesQuery(QtWidgets.QDialog):
@@ -209,477 +121,6 @@ class OverwriteFilesQuery(QtWidgets.QDialog):
 
     def set_cancel(self):
         self.choice = "cancel"
-
-
-class NetworkListItem(QtWidgets.QListWidgetItem):
-
-    def __init__(self, parent=None):
-        super(NetworkListItem, self).__init__(parent)
-        self.ntwk = None
-        self.ntwk_corrected = None
-        self.parameters = {}
-
-    def update_ntwk_names(self, name):
-        if isinstance(self.ntwk, skrf.Network):
-            self.ntwk.name = name
-        if isinstance(self.ntwk_corrected, skrf.Network):
-            self.ntwk_corrected.name = name + "-cal"
-
-    def set_text(self):
-        self.setText(self.ntwk.name)
-
-
-class NetworkListWidget(QtWidgets.QListWidget):
-    item_removed = QtCore.Signal()
-    item_updated = QtCore.Signal(object)
-    save_single_requested = QtCore.Signal(object, str)
-
-    def __init__(self, name_prefix='meas', parent=None):
-        super(NetworkListWidget, self).__init__(parent)
-        self.name_prefix = name_prefix
-        self.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-        self.setSelectionMode(QtWidgets.QAbstractItemView.ExtendedSelection)
-
-        self.customContextMenuRequested.connect(self.list_item_right_clicked)
-        self.save_single_requested.connect(save_NetworkListItem)
-        self.itemDelegate().commitData.connect(self.item_text_updated)
-        self.itemSelectionChanged.connect(self.set_active_networks)
-        self.itemClicked.connect(self.set_active_networks)
-        self.item_updated.connect(self.set_active_network)
-
-        self._ntwk_plot = None
-        self.named_items = {}
-
-    @property
-    def ntwk_plot(self):
-        return self._ntwk_plot  # type: NetworkPlotWidget
-
-    @ntwk_plot.setter
-    def ntwk_plot(self, ntwk_plot):
-        if isinstance(ntwk_plot, NetworkPlotWidget):
-            self._ntwk_plot = ntwk_plot
-            self.item_removed.connect(self._ntwk_plot.clear_plot)
-        else:
-            self._ntwk_plot = None
-            self.item_removed.disconnect()
-
-    def get_unique_name(self, name=None, exclude_item=-1):
-        """
-        :type name: str
-        :type exclude_item: int
-        :return:
-        """
-        names = []
-        for i in range(self.count()):
-            if i == exclude_item:
-                continue
-
-            item = self.item(i)
-            names.append(item.ntwk.name)
-
-        if name in names:
-            if re.match("_\d\d", name[-3:]):
-                name_base = name[:-3]
-                suffix = int(name[-2:])
-            else:
-                name_base = name
-                suffix = 1
-
-            for num in range(suffix, 100, 1):
-                name = "{:s}_{:02d}".format(name_base, num)
-                if name not in names:
-                    break
-        return name
-
-    def item_text_updated(self, emit=True):
-        item = self.currentItem()  # type: NetworkListItem
-        item.setText(self.get_unique_name(item.ntwk.name, self.row(item)))
-        item.update_ntwk_names(item.text())
-        if emit:
-            self.item_updated.emit(item)
-
-    def list_item_right_clicked(self, position):
-        menu = QtWidgets.QMenu()
-
-        if len(self.selectedItems()) == 1:
-            save = QtWidgets.QAction("Save Item", self)
-            menu.addAction(save)
-            save.triggered.connect(self.save_selected_items)
-
-            remove = QtWidgets.QAction("Remove Item", self)
-            menu.addAction(remove)
-            remove.triggered.connect(self.remove_item)
-        elif len(self.selectedItems()) > 1:
-            save = QtWidgets.QAction("Save Items", self)
-            menu.addAction(save)
-            save.triggered.connect(self.save_selected_items)
-
-            remove = QtWidgets.QAction("Remove Items", self)
-            menu.addAction(remove)
-            remove.triggered.connect(self.remove_item)
-
-        menu.exec_(self.mapToGlobal(position))  # QtWidgets.QAction
-
-    def remove_item(self):
-        items = self.selectedItems()
-        if len(items) > 0:
-            self.item_removed.emit()
-            for item in items:
-                self.takeItem(self.row(item))
-
-    def get_save_which_mode(self):
-        """
-        because the item will potentially have a raw and calibrated network attached we need
-        to determine if we want to save
-        "raw", "cal", or "both"
-        the default will be to save raw, and this method must be replaced by the parent infrastructure
-        for a different result
-
-        :return: str
-        """
-        return "raw"
-
-    def set_active_networks(self):
-        if not self.ntwk_plot:
-            return
-
-        items = self.selectedItems()        
-        ntwk_list = []
-        ntwk_list_corrected = []
-
-        for item in items:
-            if isinstance(item.ntwk, skrf.Network):
-                ntwk_list.append(item.ntwk)
-            elif type(item.ntwk) in (list, tuple):
-                ntwk_list.extend(item.ntwk)
-                
-            if isinstance(item.ntwk_corrected, skrf.Network):
-                ntwk_list_corrected.append(item.ntwk_corrected)
-            elif type(item.ntwk_corrected) in (list, tuple):
-                ntwk_list_corrected.extend(item.ntwk_corrected)
-        
-        if ntwk_list:
-            ntwk = ntwk_list if len(ntwk_list) > 1 else ntwk_list[0]
-        else:
-            ntwk = None
-
-        if ntwk_list_corrected:
-            ntwk_corrected = ntwk_list_corrected if len(ntwk_list_corrected) > 1 else ntwk_list_corrected[0]
-        else:
-            ntwk_corrected = None
-
-        self.ntwk_plot.set_networks(ntwk, ntwk_corrected)
-
-    def set_active_network(self, item):
-        """
-        :type item: NetworkListItem
-        :return:
-        """
-        self.blockSignals(True)
-        if item is None:
-            return
-
-        if self.ntwk_plot:
-            self.ntwk_plot.set_networks(item.ntwk, item.ntwk_corrected)
-        self.blockSignals(False)
-
-    def load_named_ntwk(self, ntwk, name, activate=True):
-        item = self.get_named_item(name)
-        if not item:
-            item = NetworkListItem()
-            self.addItem(item)
-
-        item.setFlags(item.flags() | QtCore.Qt.NoItemFlags)
-        item.setText(name)
-        item.ntwk = ntwk
-        item.update_ntwk_names(name)
-
-        self.clearSelection()
-        self.setCurrentItem(item)
-        if activate is True:
-            self.set_active_network(item)
-
-    def get_list_of_names(self):
-        names = list()
-        for i in range(self.count()):
-            item = self.item(i)
-            names.append(item.ntwk.name)
-        return names
-
-    def get_named_item(self, name):
-        named_item = None
-        print(name)
-        for i in range(self.count()):
-            item = self.item(i)
-            print(item.ntwk.name)
-            if item.ntwk.name == name:
-                named_item = item
-        return named_item
-
-    def load_network(self, ntwk, activate=True):
-        item = NetworkListItem()
-        item.setFlags(item.flags() | QtCore.Qt.ItemIsEditable)
-        item.setText(ntwk.name)
-        item.ntwk = ntwk
-        self.addItem(item)
-        self.clearSelection()
-        self.setCurrentItem(item)
-        self.item_text_updated(emit=activate)
-        return item
-
-    def load_from_file(self, caption="load touchstone file"):
-        ntwk = load_network_file(caption)  # type: skrf.Network
-        if not ntwk:
-            return
-        self.load_network(ntwk)
-
-    def load_networks(self, ntwks):
-        if not ntwks:
-            return
-
-        if isinstance(ntwks, skrf.Network):
-            self.load_network(ntwks)
-            return
-
-        try:
-            self.blockSignals(False)
-            for ntwk in ntwks:
-                self.load_network(ntwk, False)
-        finally:
-            self.blockSignals(False)
-
-        item = self.item(self.count() - 1)
-        self.set_active_network(item)
-
-    def load_from_files(self, caption="load touchstone file"):
-        self.load_networks(load_network_files(caption))
-
-    def save_selected_items(self):
-        items = self.selectedItems()
-        if len(items) == 1:
-            item = items[0]
-            self.save_single_requested.emit(item, self.get_save_which_mode())
-        elif len(items) > 1:
-            ntwk_list = []
-            for item in items:
-                ntwk_list.append(item.ntwk)
-            save_multiple_networks(ntwk_list)
-
-    def save_all_measurements(self):
-        save_which = self.get_save_which_mode()
-        ntwk_list = []
-
-        for i in range(self.count()):
-            item = self.item(i)
-            if save_which != "cal" and isinstance(item.ntwk, skrf.Network):
-                ntwk_list.append(item.ntwk)
-            if save_which != "raw" and isinstance(item.ntwk_corrected, skrf.Network):
-                ntwk_list.append(item.ntwk_corrected)
-
-        save_multiple_networks(ntwk_list)
-
-    def get_analyzer(self):
-        raise AttributeError("Must set get_analyzer method externally")
-
-    def get_all_networks(self, corrected=False):
-        ntwks = []
-        for i in range(self.count()):
-            item = self.item(i)
-            ntwk = item.ntwk_corrected if corrected is True else item.ntwk
-            if isinstance(ntwk, skrf.Network):
-                ntwks.append(ntwk)
-            elif type(ntwk) in (list, tuple):
-                ntwks.extend(ntwk)
-        return ntwks
-
-    def measure_ntwk(self):
-        with self.get_analyzer() as nwa:
-            dialog = MeasurementDialog(nwa)
-            dialog.measurements_available.connect(self.load_networks)
-            dialog.exec_()
-        #     meas = nwa.measure_twoport_ntwk()
-        #     meas.name = self.name_prefix  # unique name processed in load_network
-        # self.load_network(meas)
-
-    def get_load_button(self, label="Load"):
-        button = QtWidgets.QPushButton(label)
-        button.released.connect(self.load_from_files)
-        return button
-
-    def get_measure_button(self, label="Measure"):
-        button = QtWidgets.QPushButton(label)
-        button.released.connect(self.measure_ntwk)
-        return button
-
-    def get_save_selected_button(self, label="Save Selected"):
-        button = QtWidgets.QPushButton(label)
-        button.clicked.connect(self.save_selected_items)
-        return button
-
-    def get_save_all_button(self, label="Save All"):
-        button = QtWidgets.QPushButton(label)
-        button.clicked.connect(self.save_all_measurements)
-        return button
-
-    def get_input_buttons(self, labels=("Load", "Measure")):
-        widget = QtWidgets.QWidget()
-        horizontal_layout = QtWidgets.QHBoxLayout(widget)
-        horizontal_layout.setContentsMargins(0, 2, 0, 2)
-        load_button = self.get_load_button(labels[0])
-        measurement_button = self.get_measure_button(labels[1])
-        horizontal_layout.addWidget(load_button)
-        horizontal_layout.addWidget(measurement_button)
-        return widget
-
-    def get_save_buttons(self, labels=("Save Selected", "Save All")):
-        widget = QtWidgets.QWidget()
-        horizontal_layout = QtWidgets.QHBoxLayout(widget)
-        horizontal_layout.setContentsMargins(0, 2, 0, 2)
-        save_selected_button = self.get_save_selected_button(labels[0])
-        save_all_button = self.get_save_all_button(labels[1])
-        horizontal_layout.addWidget(save_selected_button)
-        horizontal_layout.addWidget(save_all_button)
-        return widget
-
-
-class ParameterizedNetworkListWidget(NetworkListWidget):
-    def __init__(self, name_prefix=None, item_parameters=(), parent=None):
-        """
-        initialize a parameterized version of the NetworkListWidget
-
-        Parameters
-        ----------
-        name_prefix : str
-            typically "meas", but some text that will be the default name of new measurements
-        item_parameters : Iterable
-            a list of dictionaries that contain the attributes each item will ahve
-        parent : QtWidgets.QWidget
-            the parent widget of the NetworkListWidget
-        """
-        super(ParameterizedNetworkListWidget, self).__init__(name_prefix, parent)
-        self._item_parameters = OrderedDict()
-        self.itemDoubleClicked.connect(self.edit_item)
-        self._label_parameters = list()
-        for param in item_parameters:
-            self.add_item_parameter(param)
-
-    @property
-    def label_parameters(self):
-        return self._label_parameters
-
-    @label_parameters.setter
-    def label_parameters(self, parameters):
-        if type(parameters) is str:
-            if parameters in self.item_parameters.keys():
-                self._label_parameters = [parameters]
-            else:
-                print("invalid parameter")
-                return
-        elif type(parameters) not in (list, tuple):
-            Warning("must provide a list of strings that identify parameters to include in the label")
-            return
-
-        label_parameters = list()
-
-        for param_name in parameters:
-            if type(param_name) is not str:
-                print("must provide strings only")
-            elif param_name in self.item_parameters.keys() and param_name not in self._label_parameters:
-                label_parameters.append(param_name)
-
-        if not label_parameters:
-            Warning("no valid parameters provided")
-        else:
-            self._label_parameters = label_parameters
-
-    def set_item_text(self, item):
-        name = item.ntwk.name
-        for param in self.label_parameters:
-            name += " - {:}".format(item.parameters[param])
-            units = self.item_parameters[param]["units"]
-            if units:
-                name += " {:}".format(units)
-        item.setText(name)
-
-    def load_network(self, ntwk, activate=True):
-        item = NetworkListItem()
-        item.setFlags(item.flags() | QtCore.Qt.NoItemFlags)
-        item.ntwk = ntwk
-        for name, param in self.item_parameters.items():
-            item.parameters[name] = param["default"]
-        self.addItem(item)
-        self.clearSelection()
-        self.setCurrentItem(item)
-        self.item_text_updated(emit=activate)
-        self.set_item_text(item)
-        return item
-
-    def edit_item(self, item):
-        """
-        pop up a dialog to edit the name and network parameters
-
-        Parameters
-        ----------
-        item : NetworkListItem
-        """
-        dialog = NetworkParameterEditor(item, self.item_parameters)
-
-        accepted = dialog.exec_()
-        if accepted:
-            for name, input in dialog.inputs.items():
-                if isinstance(input, numeric_inputs.NumericLineEdit):
-                    value = input.get_value()
-                elif isinstance(input, QtWidgets.QLineEdit):
-                    value = input.text()
-                elif isinstance(input, QtWidgets.QComboBox):
-                    value = input.currentText()
-                elif isinstance(input, QtWidgets.QDoubleSpinBox) or isinstance(input, QtWidgets.QSpinBox):
-                    value = input.value()
-                else:
-                    raise TypeError("Unsupported Widget {:}".format(input))
-                item.parameters[name] = value
-
-            index = self.indexFromItem(item).row()
-            name = self.get_unique_name(dialog.inputs["name"].text(), index)
-            item.update_ntwk_names(name)
-            self.set_item_text(item)
-
-    @property
-    def item_parameters(self):
-        return self._item_parameters
-
-    def add_item_parameter(self, attribute):
-        name = attribute["name"]
-        attr_type = attribute["type"]
-        default = attribute["default"]
-        units = attribute.get("units", None)
-        combo_list = attribute.get("combo_list", None)
-
-        self._item_parameters[name] = {
-            "type": attr_type,
-            "default": default,
-            "units": units,
-            "combo_list": combo_list
-        }
-
-    def get_parameter_from_all(self, parameter):
-        """
-        retrieve the value of a parameter from the all of the items, and return as a list
-
-        Parameters
-        ----------
-        parameter : str
-            the dict key of the parameter of interest
-
-        Returns
-        -------
-        list
-        """
-        values = list()
-        for i in range(self.count()):
-            item = self.item(i)
-            values.append(item.parameters[parameter])
-        return values
 
 
 class NetworkParameterEditor(QtWidgets.QDialog):
@@ -733,398 +174,6 @@ class NetworkParameterEditor(QtWidgets.QDialog):
         cp = QtWidgets.QDesktopWidget().availableGeometry().center()
         qr.moveCenter(cp)
         self.move(qr.topLeft())
-
-
-class NetworkPlotWidget(QtWidgets.QWidget):
-    S_VALS = OrderedDict((
-        ("decibels", "db"),
-        ("magnitude", "mag"),
-        ("phase (deg)", "deg"),
-        ("phase unwrapped (deg)", "deg_unwrap"),
-        ("phase (rad)", "rad"),
-        ("phase unwrapped (rad)", "rad_unwrap"),
-        ("real", "re"),
-        ("imaginary", "im"),
-    ))
-    S_UNITS = list(S_VALS.keys())
-
-    def __init__(self, parent=None, **kwargs):
-        super(NetworkPlotWidget, self).__init__(parent)
-
-        self.checkBox_useCorrected = QtWidgets.QCheckBox()
-        self.checkBox_useCorrected.setText("Plot Corrected")
-
-        self.comboBox_primarySelector = QtWidgets.QComboBox(self)
-        self.comboBox_primarySelector.addItems(("S", "Z", "Y", "A", "Smith Chart"))
-
-        self.comboBox_unitsSelector = QtWidgets.QComboBox(self)
-        self.comboBox_unitsSelector.addItems(self.S_UNITS)
-
-        self.comboBox_traceSelector = QtWidgets.QComboBox(self)
-        self.set_trace_items()
-        self.comboBox_traceSelector.setCurrentIndex(0)
-
-        self.plot_layout = pg.GraphicsLayoutWidget(self)
-        self.plot_layout.sceneObj.sigMouseClicked.connect(self.graph_clicked)
-
-        self.horizontalLayout = QtWidgets.QHBoxLayout()
-        self.horizontalLayout.addWidget(self.checkBox_useCorrected)
-        self.horizontalLayout.addWidget(self.comboBox_primarySelector)
-        self.horizontalLayout.addWidget(self.comboBox_unitsSelector)
-        self.horizontalLayout.addWidget(self.comboBox_traceSelector)
-
-        self.data_info_label = QtWidgets.QLabel("Click a data point to see info")
-
-        self.verticalLayout = QtWidgets.QVBoxLayout(self)
-        self.verticalLayout.setContentsMargins(3, 3, 3, 3)  # normally this will be embedded in another application
-        self.verticalLayout.addLayout(self.horizontalLayout)
-        self.verticalLayout.addWidget(self.plot_layout)
-        self.verticalLayout.addWidget(self.data_info_label)
-
-        self.checkBox_useCorrected.stateChanged.connect(self.set_use_corrected)
-        self.comboBox_primarySelector.currentIndexChanged.connect(self.update_plot)
-        self.comboBox_unitsSelector.currentIndexChanged.connect(self.update_plot)
-        self.comboBox_traceSelector.currentIndexChanged.connect(self.update_plot)
-        
-        self.plot = self.plot_layout.addPlot()  # type: pg.PlotItem
-
-        self._ntwk = None
-        self._ntwk_corrected = None
-        self._corrected_data_enabled = True
-        self._use_corrected = False
-        self.corrected_data_enabled = kwargs.get('corrected_data_enabled', True)
-
-        self.plot.addLegend()
-        self.plot.showGrid(True, True)
-        self.plot.setLabel("bottom", "frequency", units="Hz")
-
-        self.last_plot = "rectangular"
-
-    def get_use_corrected(self):
-        return self._use_corrected
-
-    def set_use_corrected(self, val):
-        if val in (1, 2):
-            self._use_corrected = True
-        else:
-            self._use_corrected = False
-        self.update_plot()
-
-    use_corrected = property(get_use_corrected, set_use_corrected)
-
-    @property
-    def ntwk(self): return self._ntwk
-
-    @ntwk.setter
-    def ntwk(self, ntwk):
-        if ntwk is None or isinstance(ntwk, skrf.Network) or type(ntwk) in (list, tuple):
-            self.set_trace_items(ntwk)
-            self._ntwk = ntwk
-            self.update_plot()
-        else:
-            raise TypeError("must set to skrf.Network, list of Networks, or None")
-
-    @property
-    def ntwk_corrected(self): return self._ntwk_corrected
-
-    @ntwk_corrected.setter
-    def ntwk_corrected(self, ntwk):
-        if ntwk is None or isinstance(ntwk, skrf.Network) or type(ntwk) in (list, tuple):
-            self.set_trace_items(ntwk)
-            self._ntwk_corrected = ntwk   
-            self.update_plot()
-        else:
-            raise TypeError("must set to skrf.Network, list of Networks, or None")
-
-    @property
-    def corrected_data_enabled(self):
-        return self._corrected_data_enabled
-
-    @corrected_data_enabled.setter
-    def corrected_data_enabled(self, enabled):
-        if enabled is True:
-            self._corrected_data_enabled = True
-            self.checkBox_useCorrected.setEnabled(True)
-        else:
-            self._corrected_data_enabled = False
-            self._use_corrected = False
-            self.checkBox_useCorrected.setEnabled(False)
-
-    def set_networks(self, ntwk, ntwk_corrected=None):
-        if ntwk is None or isinstance(ntwk, skrf.Network) or type(ntwk) in (list, tuple):
-            self._ntwk = ntwk
-            self.set_trace_items(self._ntwk)
-            if ntwk is None:
-                self._ntwk_corrected = None
-                self.set_trace_items(self._ntwk)
-                return
-        else:
-            raise TypeError("must set to skrf.Network, list of Networks, or None")
-
-        if ntwk_corrected is None or isinstance(ntwk_corrected, skrf.Network) or type(ntwk_corrected) in (list, tuple):
-            self._ntwk_corrected = ntwk_corrected
-        else:
-            raise TypeError("must set to skrf.Network, list of Networks, or None")
-
-        self.update_plot()
-
-    def reset_plot(self, smith=False):
-        self.plot.clear()
-
-        if not smith and self.last_plot == "smith":
-            self.plot.setAspectLocked(False)
-            self.plot.autoRange()
-            self.plot.enableAutoRange()
-            self.plot.setLabel("bottom", "frequency", units="Hz")
-
-        if smith and not self.last_plot == "smith":
-            self.last_plot = "smith"
-            self.ZGrid = smith_chart.gen_z_grid()
-            self.s_unity_circle = smith_chart.gen_s_unity_circle()
-            self.plot_layout.removeItem(self.plot)
-            self.plot = self.plot_layout.addPlot()
-            self.plot.setAspectLocked()
-            self.plot.setXRange(-1, 1)
-            self.plot.setYRange(-1, 1)
-
-        if smith:
-            self.plot.addItem(self.s_unity_circle)
-            self.plot.addItem(self.ZGrid)
-
-        if not smith:
-            self.plot.setLabel("left", "")
-
-        self.plot.setTitle(None)
-        legend = self.plot.legend
-        if legend is not None:
-            legend.scene().removeItem(legend)
-        self.plot.addLegend()
-
-    def clear_plot(self):
-        self._ntwk = None
-        self._ntwk_corrected = None
-        self._ntwk_list = None
-        self.reset_plot()
-
-    def set_trace_items(self, ntwk=None):
-        self.comboBox_traceSelector.blockSignals(True)
-        current_index = self.comboBox_traceSelector.currentIndex()
-        nports = 0
-
-        if isinstance(ntwk, skrf.Network):
-            nports = ntwk.nports
-        elif type(ntwk) in (list, tuple):
-            for n in ntwk:
-                if n.nports > nports:
-                    nports = n.nports
-
-        self.comboBox_traceSelector.clear()
-        self.comboBox_traceSelector.addItem("all")
-
-        for n in range(nports):
-            for m in range(nports):
-                self.comboBox_traceSelector.addItem("S{:d}{:d}".format(m + 1, n + 1))
-
-        if current_index <= self.comboBox_traceSelector.count():
-            self.comboBox_traceSelector.setCurrentIndex(current_index)
-        else:
-            self.comboBox_traceSelector.setCurrentIndex(0)
-        self.comboBox_traceSelector.blockSignals(False)
-
-    def graph_clicked(self, ev):
-        """
-        :type ev: pg.GraphicsScene.mouseEvents.MouseClickEvent
-        :return:
-        """
-        xy = self.plot.vb.mapSceneToView(ev.scenePos())
-        if not ev.isAccepted():
-            if "smith" in self.comboBox_primarySelector.currentText().lower():
-                S11 = xy.x() + 1j * xy.y()
-                Z = (1 + S11) / (1 - S11)
-                self.data_info_label.setText(
-                    "Sre: {:g}, Sim: {:g}  -  R: {:g}, X: {:g}".format(xy.x(), xy.y(), Z.real, Z.imag))
-            else:
-                self.data_info_label.setText("x: {:g}, y: {:g}".format(xy.x(), xy.y()))
-        elif isinstance(ev.acceptedItem, pg.PlotCurveItem):
-            curve = ev.acceptedItem  # type: pg.PlotCurveItem
-            spoint = xy.x() + 1j * xy.y()
-            sdata = curve.xData + 1j * curve.yData
-            index = np.argmin(np.abs(sdata - spoint))
-            frequency = curve.ntwk.frequency.f_scaled[index]
-            S11 = curve.xData[index] + 1j * curve.yData[index]
-            Z = (1 + S11) / (1 - S11)
-            self.data_info_label.setText(
-                "Freq: {:g} ({:s}), S(re): {:g}, S(im): {:g}  -  R: {:g}, X: {:g}".format(
-                    frequency, curve.ntwk.frequency.unit, S11.real, S11.imag, Z.real, Z.imag))
-
-    def update_plot(self):
-        if "smith" in self.comboBox_primarySelector.currentText().lower():
-            self.plot_smith()
-        else:
-            self.plot_ntwk()
-            self.last_plot = "rectangular"
-
-    def plot_ntwk(self):
-        import time
-        print("plotting", time.time())
-        if self.use_corrected and self.ntwk_corrected is not None:
-            ntwk = self.ntwk_corrected
-        else:
-            ntwk = self.ntwk
-
-        if ntwk is None:
-            return
-        elif type(ntwk) in (list, tuple):
-            self.plot_ntwk_list()
-            return
-
-        self.reset_plot()
-        self.plot.showGrid(True, True)
-        self.plot.setLabel("bottom", "frequency", units="Hz")
-
-        colors = trace_color_cycle(ntwk.s.shape[1] ** 2)
-
-        trace = self.comboBox_traceSelector.currentIndex()
-        n_ = m_ = 0
-        if trace > 0:
-            mn = trace - 1
-            nports = int(sqrt(self.comboBox_traceSelector.count() - 1))
-            m_ = mn % nports
-            n_ = int((mn - mn % nports) / nports)
-
-        primary = self.comboBox_primarySelector.currentText().lower()
-        s_units = self.comboBox_unitsSelector.currentText()
-        attr = primary + "_" + self.S_VALS[s_units]
-        s = getattr(ntwk, attr)
-        for n in range(ntwk.s.shape[2]):
-            for m in range(ntwk.s.shape[1]):
-                c = next(colors)
-
-                if trace > 0:
-                    if not n == n_ or not m == m_:
-                        continue
-
-                label = "S{:d}{:d}".format(m + 1, n + 1)
-
-                self.plot.plot(ntwk.f, s[:, m, n], pen=pg.mkPen(c), name=label)
-        self.plot.setLabel("left", s_units)
-        self.plot.setTitle(ntwk.name)
-
-    def plot_ntwk_list(self):
-        ntwk_list = self.ntwk
-        if ntwk_list is None:
-            return
-
-        self.reset_plot()
-        self.plot.showGrid(True, True)
-        self.plot.setLabel("bottom", "frequency", units="Hz")
-
-        colors = trace_color_cycle()
-
-        trace = self.comboBox_traceSelector.currentIndex()
-        n_ = m_ = 0
-        if trace > 0:
-            mn = trace - 1
-            nports = int(sqrt(self.comboBox_traceSelector.count() - 1))
-            m_ = mn % nports
-            n_ = int((mn - mn % nports) / nports)
-
-        primary = self.comboBox_primarySelector.currentText().lower()
-        s_units = self.comboBox_unitsSelector.currentText()
-        attr = primary + "_" + self.S_VALS[s_units]
-
-        for ntwk in ntwk_list:
-            s = getattr(ntwk, attr)
-            for n in range(ntwk.s.shape[2]):
-                for m in range(ntwk.s.shape[1]):
-                    c = next(colors)
-
-                    if trace > 0:
-                        if not n == n_ or not m == m_:
-                            continue
-
-                    label = ntwk.name
-                    if ntwk.s.shape[1] > 1:
-                        label += " - S{:d}{:d}".format(m + 1, n + 1)
-
-                    self.plot.plot(ntwk.f, s[:, m, n], pen=pg.mkPen(c), name=label)
-        self.plot.setLabel("left", s_units)
-
-    def plot_smith(self):
-        if self.use_corrected and self.ntwk_corrected is not None:
-            ntwk = self.ntwk_corrected
-        else:
-            ntwk = self.ntwk
-
-        if ntwk is None:
-            self.reset_plot(smith=True)
-            return
-        elif type(ntwk) in (list, tuple):
-            self.plot_smith_list()
-            return
-
-        self.reset_plot(smith=True)
-
-        colors = trace_color_cycle(ntwk.s.shape[1] ** 2)
-
-        trace = self.comboBox_traceSelector.currentIndex()
-        n_ = m_ = 0
-        if trace > 0:
-            mn = trace - 1
-            nports = int(sqrt(self.comboBox_traceSelector.count() - 1))
-            m_ = mn % nports
-            n_ = int((mn - mn % nports) / nports)
-
-        for n in range(ntwk.s.shape[2]):
-            for m in range(ntwk.s.shape[1]):
-                c = next(colors)
-
-                if trace > 0:
-                    if not n == n_ or not m == m_:
-                        continue
-
-                label = "S{:d}{:d}".format(m + 1, n + 1)
-
-                s = ntwk.s[:, m, n]
-                curve = self.plot.plot(s.real, s.imag, pen=pg.mkPen(c), name=label)
-                curve.curve.setClickable(True)
-                curve.curve.ntwk = ntwk
-        self.plot.setTitle(ntwk.name)
-
-    def plot_smith_list(self):
-        self.reset_plot(smith=True)
-
-        ntwk_list = self.ntwk
-        if ntwk_list is None:
-            return
-
-        colors = trace_color_cycle()
-
-        trace = self.comboBox_traceSelector.currentIndex()
-        n_ = m_ = 0
-        if trace > 0:
-            mn = trace - 1
-            nports = int(sqrt(self.comboBox_traceSelector.count() - 1))
-            m_ = mn % nports
-            n_ = int((mn - mn % nports) / nports)
-
-        for ntwk in ntwk_list:
-            for n in range(ntwk.s.shape[2]):
-                for m in range(ntwk.s.shape[1]):
-                    c = next(colors)
-
-                    if trace > 0:
-                        if not n == n_ or not m == m_:
-                            continue
-
-                    label = ntwk.name
-                    if ntwk.s.shape[1] > 1:
-                        label += " - S{:d}{:d}".format(m + 1, n + 1)
-
-                    s = ntwk.s[:, m, n]
-                    curve = self.plot.plot(s.real, s.imag, pen=pg.mkPen(c), name=label)
-                    curve.curve.setClickable(True)
-                    curve.curve.ntwk = ntwk
 
 
 class SwitchTermsDialog(QtWidgets.QDialog):
@@ -1458,9 +507,10 @@ class ReflectDialog(QtWidgets.QDialog):
             self.label_port2.setText("port2 - measured")
         else:
             if type(self.s11) is skrf.Network and type(self.s22) is skrf.Network:
-                self.reflect_2port = skrf.four_oneports_2_twoport(self.s11, self.s11, self.s22, self.s22)
-                self.reflect_2port.s[:, 0, 1] = 0
-                self.reflect_2port.s[:, 1, 0] = 0
+                self.reflect_2port = skrf.two_port_reflect(self.s11, self.s22)
+                # self.reflect_2port = skrf.four_oneports_2_twoport(self.s11, self.s11, self.s22, self.s22)
+                # self.reflect_2port.s[:, 0, 1] = 0
+                # self.reflect_2port.s[:, 1, 0] = 0
                 self.ready = True
                 self.label_port1.setText("port1 - measured")
                 self.label_port2.setText("port2 - measured")
